refactor: :art: moves function to component in a separate file
